+++ make_label.py
@@ -4,7 +4,7 @@ from train import train
 
 def make_label(file):
     csv_int = [list(map(int,line.rstrip().split(","))) for line in open(file).readlines()]
-    csv_torch = torch.tensor(csv_int).long()
+    csv_torch = torch.tensor(csv_int).float()
     return csv_torch
 
 # train_label = make_label('train_label.csv').float()
